chore: update convert.py
